requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
pdfplumber
google-genai
python-dotenv
numpy
chromadb
pydantic-settings